Split browser stuff out of utlis to rm shim dep
